+++ implement/Pine.Core.Tests/Elm/ElmCompilerInDotnet/SequentialIREfficiencyTests.cs
@@ -298,7 +298,7 @@ public class SequentialIREfficiencyTests
 
         renderedFrame.Should().Be(
             """
-            02085566 (9):
+            658be8d6 (9):
             0: Local_Get_Skip_Head_Const (0, 2)
             1: Take_Const (4)
             2: Local_Set (1)
@@ -307,11 +307,11 @@ public class SequentialIREfficiencyTests
               case Blob [4] (0x00000042 | UTF32 "B"): jump (4, 7)
             4: Local_Get (1)
             5: Build_List_With_Prefix (2, 1)
-              Blob [52] (0x0000003c00000043000000680000006f0000006900000063000000650000005f... | UTF32 "\u003CChoice_Type\u003E")
+              Blob [32] (0x0000003c00000043000000680000006f0000006900000063000000650000003e | UTF32 "\u003CChoice\u003E")
               Blob [24] (0x000000530000007400000072000000690000006e00000067 | UTF32 "String")
             6: Return
             jumps_arriving_from 1 (3)
-            7: Push_Literal (List [3] (3 | hash 0xc0894db7))
+            7: Push_Literal (List [3] (3 | hash 0xe24c06b3))
             8: Return
             """);
     }
@@ -447,7 +447,7 @@ public class SequentialIREfficiencyTests
 
         renderedFrame.Should().Be(
             """
-            bacd3d53 (8):
+            0dbec914 (8):
             0: Local_Get (2)
             1: Local_Get (1)
             2: Int_Mul_Const (4)
@@ -455,7 +455,7 @@ public class SequentialIREfficiencyTests
             4: Int_Mul_Const (4)
             5: Slice_Skip_Var_Take_Var
             6: Build_List_With_Prefix (2, 1)
-              Blob [52] (0x0000003c00000043000000680000006f0000006900000063000000650000005f... | UTF32 "\u003CChoice_Type\u003E")
+              Blob [32] (0x0000003c00000043000000680000006f0000006900000063000000650000003e | UTF32 "\u003CChoice\u003E")
               Blob [24] (0x000000530000007400000072000000690000006e00000067 | UTF32 "String")
             7: Return
             """);
@@ -496,13 +496,13 @@ public class SequentialIREfficiencyTests
 
         renderedFrame.Should().Be(
             """
-            081793dc (6):
+            8f52db35 (6):
             0: Local_Get (1)
             1: Local_Get (0)
             2: Int_Mul_Const (4)
             3: Slice_Skip_Var_Take_Const (12)
             4: Build_List_With_Prefix (2, 1)
-              Blob [52] (0x0000003c00000043000000680000006f0000006900000063000000650000005f... | UTF32 "\u003CChoice_Type\u003E")
+              Blob [32] (0x0000003c00000043000000680000006f0000006900000063000000650000003e | UTF32 "\u003CChoice\u003E")
               Blob [24] (0x000000530000007400000072000000690000006e00000067 | UTF32 "String")
             5: Return
             """);