+++ implement/Pine.Core/Elm/ElmCompilerInDotnet/PrecompiledLeaves/CoreStringPrecompiledLeaves.cs
@@ -604,7 +604,7 @@ public static class CoreStringPrecompiledLeaves
     {
         if (value is PineValue.ListValue stringValue &&
             stringValue.Items.Length is 3 &&
-            stringValue.Items.Span[0] == ElmValue.ElmChoiceTypeTagNameAsValue &&
+            ElmValue.IsFlatChoiceTypeTag(stringValue.Items.Span[0]) &&
             stringValue.Items.Span[1] == ElmValue.ElmStringTypeTagNameAsValue &&
             stringValue.Items.Span[2] is PineValue.BlobValue charsBlob &&
             charsBlob.Bytes.Length % 4 is 0)